pyproject.toml
[tool.poetry]
name = "futebol-stats"
version = "0.1.0"
description = ""
authors = ["emersontanuri <emersontanuri1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
mage-ai = "^0.9.34"
google-cloud-bigquery = "^3.12.0"
python-dotenv = "^1.0.0"
google-cloud-storage = "^2.12.0"
pandas-gbq = "^0.19.2"
streamlit = "^1.27.2"
gcsfs = "^2023.9.2"
levenshtein = "^0.23.0"
python-slugify = "^8.0.1"
matplotlib = "^3.8.1"
seaborn = "^0.13.0"
streamlit-extras = "^0.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
